+++ internal/accrual/service/reward_service.go
@@ -5,15 +5,15 @@ import (
 	"github.com/Grifonhard/Practicum-s5_6/internal/lib/math"
 )
 
-func CalculateReward(good model.Good, accrual model.AccrualProgram) uint64 {
-	var result uint64
+func CalculateReward(good model.Good, accrual model.AccrualProgram) float64 {
+	var result float64
 
 	switch accrual.RewardType {
 	case model.RewardTypePoints:
-		result = uint64(accrual.Reward)
+		result = float64(accrual.Reward)
 	case model.RewardTypePercent:
 		reward := math.Percent(int(accrual.Reward), int(good.Price))
-		result = uint64(reward)
+		result = float64(reward)
 	}
 
 	return result
